Move credentials from config.json to .env
Necessary because we'll need a .env file for DB credentials in Docker
(can't easily read JSON into environment variables in docker-compose).
Because we want everything the user has to set up in one place, we
might as well move the credentials from config.json into the new .env,
then config.json won't need to be a secret anymore.