Ensure connection to MongoDB is established
This is necessary because the MongoDB driver will throw an exception
when it's unable to connect to the database on the first try.
This happens when MongoDB is a bit slower to start up than our
application, even though the app is started after the container for
MongoDB has been created. This is an issue that needs to be solved on
the application side, which is exactly what we do here.