Implement POC for server-side websocket
Clients can connect to ws://{host}:8080 and will receive a JSON object
whenever a new measurement has come in over MQTT. The objects are the
same as you'd receive inside the array returned by the API, with the
exception of an added property 'kind'. It contains the name of the
metric the object contains, so the client can simply check this property
to know which properties it can access, without inefficiently checking
for the existence of a property to determine the kind of metric.